moves user account service integration tests to new integration test folder
