+++ backend/services/mgmt/v1alpha1/integration_tests/user-account-service_integration_test.go
@@ -1,0 +1,353 @@
+package integrationtests_test
+
+import (
+	"connectrpc.com/connect"
+	"github.com/google/uuid"
+	mgmtv1alpha1 "github.com/nucleuscloud/neosync/backend/gen/go/protos/mgmt/v1alpha1"
+	"github.com/nucleuscloud/neosync/backend/gen/go/protos/mgmt/v1alpha1/mgmtv1alpha1connect"
+	"github.com/nucleuscloud/neosync/backend/internal/authmgmt"
+	pg_models "github.com/nucleuscloud/neosync/backend/sql/postgresql/models"
+	"github.com/stretchr/testify/mock"
+	"github.com/stretchr/testify/require"
+)
+
+func (s *IntegrationTestSuite) Test_UserAccountService_GetAccountOnboardingConfig() {
+	accountId := s.createPersonalAccount(s.unauthUserClient)
+
+	resp, err := s.unauthUserClient.GetAccountOnboardingConfig(s.ctx, connect.NewRequest(&mgmtv1alpha1.GetAccountOnboardingConfigRequest{AccountId: accountId}))
+	requireNoErrResp(s.T(), resp, err)
+	onboardingConfig := resp.Msg.GetConfig()
+	require.NotNil(s.T(), onboardingConfig)
+
+	require.False(s.T(), onboardingConfig.GetHasCreatedSourceConnection())
+	require.False(s.T(), onboardingConfig.GetHasCreatedDestinationConnection())
+	require.False(s.T(), onboardingConfig.GetHasCreatedJob())
+	require.False(s.T(), onboardingConfig.GetHasInvitedMembers())
+}
+
+func (s *IntegrationTestSuite) Test_UserAccountService_GetAccountOnboardingConfig_NoAccount() {
+	resp, err := s.unauthUserClient.GetAccountOnboardingConfig(s.ctx, connect.NewRequest(&mgmtv1alpha1.GetAccountOnboardingConfigRequest{AccountId: uuid.NewString()}))
+	requireErrResp(s.T(), resp, err)
+	requireConnectError(s.T(), err, connect.CodePermissionDenied)
+}
+
+func (s *IntegrationTestSuite) Test_UserAccountService_SetAccountOnboardingConfig_NoAccount() {
+	resp, err := s.unauthUserClient.SetAccountOnboardingConfig(s.ctx, connect.NewRequest(&mgmtv1alpha1.SetAccountOnboardingConfigRequest{AccountId: uuid.NewString(), Config: &mgmtv1alpha1.AccountOnboardingConfig{}}))
+	requireErrResp(s.T(), resp, err)
+	requireConnectError(s.T(), err, connect.CodePermissionDenied)
+}
+
+func (s *IntegrationTestSuite) Test_UserAccountService_SetAccountOnboardingConfig_NoConfig() {
+	accountId := s.createPersonalAccount(s.unauthUserClient)
+
+	resp, err := s.unauthUserClient.SetAccountOnboardingConfig(s.ctx, connect.NewRequest(&mgmtv1alpha1.SetAccountOnboardingConfigRequest{AccountId: accountId, Config: nil}))
+	requireNoErrResp(s.T(), resp, err)
+	onboardingConfig := resp.Msg.GetConfig()
+	require.NotNil(s.T(), onboardingConfig)
+
+	require.False(s.T(), onboardingConfig.GetHasCreatedSourceConnection())
+	require.False(s.T(), onboardingConfig.GetHasCreatedDestinationConnection())
+	require.False(s.T(), onboardingConfig.GetHasCreatedJob())
+	require.False(s.T(), onboardingConfig.GetHasInvitedMembers())
+}
+
+func (s *IntegrationTestSuite) Test_UserAccountService_SetAccountOnboardingConfig() {
+	accountId := s.createPersonalAccount(s.unauthUserClient)
+
+	resp, err := s.unauthUserClient.SetAccountOnboardingConfig(s.ctx, connect.NewRequest(&mgmtv1alpha1.SetAccountOnboardingConfigRequest{
+		AccountId: accountId, Config: &mgmtv1alpha1.AccountOnboardingConfig{
+			HasCreatedSourceConnection:      true,
+			HasCreatedDestinationConnection: true,
+			HasCreatedJob:                   true,
+			HasInvitedMembers:               true,
+		}},
+	))
+	requireNoErrResp(s.T(), resp, err)
+
+	onboardingConfig := resp.Msg.GetConfig()
+	require.NotNil(s.T(), onboardingConfig)
+
+	require.True(s.T(), onboardingConfig.GetHasCreatedSourceConnection())
+	require.True(s.T(), onboardingConfig.GetHasCreatedDestinationConnection())
+	require.True(s.T(), onboardingConfig.GetHasCreatedJob())
+	require.True(s.T(), onboardingConfig.GetHasInvitedMembers())
+}
+
+var (
+	validTemporalConfigModel = &pg_models.TemporalConfig{
+		Namespace:        "foo",
+		SyncJobQueueName: "bar",
+		Url:              "http://localhost:7070",
+	}
+)
+
+func (s *IntegrationTestSuite) Test_GetAccountTemporalConfig() {
+	accountId := s.createPersonalAccount(s.unauthUserClient)
+
+	s.mockTemporalClientMgr.On("GetTemporalConfigByAccount", mock.Anything, mock.Anything).
+		Return(validTemporalConfigModel, nil)
+
+	resp, err := s.unauthUserClient.GetAccountTemporalConfig(s.ctx, connect.NewRequest(&mgmtv1alpha1.GetAccountTemporalConfigRequest{AccountId: accountId}))
+	requireNoErrResp(s.T(), resp, err)
+
+	tc := resp.Msg.GetConfig()
+	require.NotNil(s.T(), tc)
+
+	require.Equal(s.T(), validTemporalConfigModel.Namespace, tc.GetNamespace())
+	require.Equal(s.T(), validTemporalConfigModel.SyncJobQueueName, tc.GetSyncJobQueueName())
+	require.Equal(s.T(), validTemporalConfigModel.Url, tc.GetUrl())
+}
+
+func (s *IntegrationTestSuite) Test_GetAccountTemporalConfig_NoAccount() {
+	resp, err := s.unauthUserClient.GetAccountTemporalConfig(s.ctx, connect.NewRequest(&mgmtv1alpha1.GetAccountTemporalConfigRequest{AccountId: uuid.NewString()}))
+	requireErrResp(s.T(), resp, err)
+	requireConnectError(s.T(), err, connect.CodePermissionDenied)
+}
+
+func (s *IntegrationTestSuite) Test_GetAccountTemporalConfig_NeosyncCloud() {
+	accountId := s.createPersonalAccount(s.unauthUserClient)
+
+	resp, err := s.ncunauthUserClient.GetAccountTemporalConfig(s.ctx, connect.NewRequest(&mgmtv1alpha1.GetAccountTemporalConfigRequest{AccountId: accountId}))
+	requireErrResp(s.T(), resp, err)
+	requireConnectError(s.T(), err, connect.CodeUnimplemented)
+}
+
+func (s *IntegrationTestSuite) Test_SetAccountTemporalConfig_NoAccount() {
+	resp, err := s.unauthUserClient.SetAccountTemporalConfig(s.ctx, connect.NewRequest(&mgmtv1alpha1.SetAccountTemporalConfigRequest{AccountId: uuid.NewString()}))
+	requireErrResp(s.T(), resp, err)
+	requireConnectError(s.T(), err, connect.CodePermissionDenied)
+}
+
+func (s *IntegrationTestSuite) Test_SetAccountTemporalConfig_NeosyncCloud() {
+	accountId := s.createPersonalAccount(s.unauthUserClient)
+
+	resp, err := s.ncunauthUserClient.SetAccountTemporalConfig(s.ctx, connect.NewRequest(&mgmtv1alpha1.SetAccountTemporalConfigRequest{AccountId: accountId}))
+	requireErrResp(s.T(), resp, err)
+	requireConnectError(s.T(), err, connect.CodeUnimplemented)
+}
+
+func (s *IntegrationTestSuite) Test_SetAccountTemporalConfig_NoConfig() {
+	accountId := s.createPersonalAccount(s.unauthUserClient)
+
+	s.mockTemporalClientMgr.On("GetTemporalConfigByAccount", mock.Anything, mock.Anything).
+		Return(validTemporalConfigModel, nil)
+
+	resp, err := s.unauthUserClient.SetAccountTemporalConfig(s.ctx, connect.NewRequest(&mgmtv1alpha1.SetAccountTemporalConfigRequest{AccountId: accountId, Config: nil}))
+	requireNoErrResp(s.T(), resp, err)
+
+	tc := resp.Msg.GetConfig()
+	require.NotNil(s.T(), tc)
+
+	require.Equal(s.T(), validTemporalConfigModel.Namespace, tc.GetNamespace())
+	require.Equal(s.T(), validTemporalConfigModel.SyncJobQueueName, tc.GetSyncJobQueueName())
+	require.Equal(s.T(), validTemporalConfigModel.Url, tc.GetUrl())
+}
+
+func (s *IntegrationTestSuite) Test_SetAccountTemporalConfig() {
+	accountId := s.createPersonalAccount(s.unauthUserClient)
+
+	// kind of a bad test since we are mocking this client wholesale, but it at least verifies we can write the config
+	s.mockTemporalClientMgr.On("GetTemporalConfigByAccount", mock.Anything, mock.Anything).
+		Return(validTemporalConfigModel, nil)
+
+	resp, err := s.unauthUserClient.SetAccountTemporalConfig(s.ctx, connect.NewRequest(&mgmtv1alpha1.SetAccountTemporalConfigRequest{
+		AccountId: accountId, Config: &mgmtv1alpha1.AccountTemporalConfig{
+			Url:              "test",
+			Namespace:        "test",
+			SyncJobQueueName: "test",
+		}}))
+	requireNoErrResp(s.T(), resp, err)
+
+	tc := resp.Msg.GetConfig()
+	require.NotNil(s.T(), tc)
+
+	require.Equal(s.T(), validTemporalConfigModel.Namespace, tc.GetNamespace())
+	require.Equal(s.T(), validTemporalConfigModel.SyncJobQueueName, tc.GetSyncJobQueueName())
+	require.Equal(s.T(), validTemporalConfigModel.Url, tc.GetUrl())
+}
+
+var (
+	validAuthUser = &authmgmt.User{Name: "foo", Email: "bar", Picture: "baz"}
+)
+
+func (s *IntegrationTestSuite) Test_GetUser_Auth() {
+	client := s.getAuthUserClient("test-user1")
+	userId := s.setUser(client)
+
+	resp, err := client.GetUser(s.ctx, connect.NewRequest(&mgmtv1alpha1.GetUserRequest{}))
+	requireNoErrResp(s.T(), resp, err)
+	require.Equal(s.T(), userId, resp.Msg.GetUserId())
+}
+
+func (s *IntegrationTestSuite) Test_GetUser_Auth_NotFound() {
+	client := s.getAuthUserClient("test-user")
+	resp, err := client.GetUser(s.ctx, connect.NewRequest(&mgmtv1alpha1.GetUserRequest{}))
+	requireErrResp(s.T(), resp, err)
+	requireConnectError(s.T(), err, connect.CodeNotFound)
+}
+
+func (s *IntegrationTestSuite) Test_SetUser_Auth() {
+	client := s.getAuthUserClient("test-user")
+	userId := s.setUser(client)
+	require.NotEmpty(s.T(), userId)
+	require.NotEqual(s.T(), "00000000-0000-0000-0000-000000000000", userId)
+}
+
+func (s *IntegrationTestSuite) Test_CreateTeamAccount_Auth() {
+	client := s.getAuthUserClient("test-user")
+	s.setUser(client)
+
+	resp, err := client.CreateTeamAccount(s.ctx, connect.NewRequest(&mgmtv1alpha1.CreateTeamAccountRequest{Name: "test-name"}))
+	requireNoErrResp(s.T(), resp, err)
+	require.NotEmpty(s.T(), resp.Msg.GetAccountId())
+}
+
+func (s *IntegrationTestSuite) Test_GetTeamAccountMembers_Auth() {
+	client := s.getAuthUserClient("test-user")
+	userId := s.setUser(client)
+	accountId := s.createTeamAccount(client, "test-team")
+
+	s.mockAuthMgmtClient.On("GetUserBySub", mock.Anything, "test-user").
+		Return(validAuthUser, nil)
+
+	resp, err := client.GetTeamAccountMembers(s.ctx, connect.NewRequest(&mgmtv1alpha1.GetTeamAccountMembersRequest{AccountId: accountId}))
+	requireNoErrResp(s.T(), resp, err)
+
+	members := resp.Msg.GetUsers()
+	require.Len(s.T(), members, 1)
+	member := members[0]
+	require.Equal(s.T(), userId, member.GetId())
+	require.Equal(s.T(), validAuthUser.Email, member.GetEmail())
+	require.Equal(s.T(), validAuthUser.Picture, member.GetImage())
+	require.Equal(s.T(), validAuthUser.Name, member.GetName())
+}
+
+func (s *IntegrationTestSuite) setUser(client mgmtv1alpha1connect.UserAccountServiceClient) string {
+	s.T().Helper()
+	resp, err := client.SetUser(s.ctx, connect.NewRequest(&mgmtv1alpha1.SetUserRequest{}))
+	requireNoErrResp(s.T(), resp, err)
+	return resp.Msg.GetUserId()
+}
+
+func (s *IntegrationTestSuite) createTeamAccount(client mgmtv1alpha1connect.UserAccountServiceClient, name string) string {
+	s.T().Helper()
+	resp, err := client.CreateTeamAccount(s.ctx, connect.NewRequest(&mgmtv1alpha1.CreateTeamAccountRequest{Name: name}))
+	requireNoErrResp(s.T(), resp, err)
+	return resp.Msg.AccountId
+}
+
+func (s *IntegrationTestSuite) Test_GetUser() {
+	resp, err := s.unauthUserClient.GetUser(s.ctx, connect.NewRequest(&mgmtv1alpha1.GetUserRequest{}))
+	requireNoErrResp(s.T(), resp, err)
+
+	userId := resp.Msg.GetUserId()
+	require.NotEmpty(s.T(), userId)
+}
+
+func (s *IntegrationTestSuite) Test_SetUser() {
+	resp, err := s.unauthUserClient.SetUser(s.ctx, connect.NewRequest(&mgmtv1alpha1.SetUserRequest{}))
+	requireNoErrResp(s.T(), resp, err)
+
+	userId := resp.Msg.UserId
+	require.NotEmpty(s.T(), userId)
+	require.Equal(s.T(), "00000000-0000-0000-0000-000000000000", userId)
+}
+
+func (s *IntegrationTestSuite) Test_GetAccounts_Empty() {
+	resp, err := s.unauthUserClient.GetUserAccounts(s.ctx, connect.NewRequest(&mgmtv1alpha1.GetUserAccountsRequest{}))
+	requireNoErrResp(s.T(), resp, err)
+
+	accounts := resp.Msg.GetAccounts()
+	require.Empty(s.T(), accounts)
+}
+
+func (s *IntegrationTestSuite) Test_SetPersonalAccount() {
+	resp, err := s.unauthUserClient.SetPersonalAccount(s.ctx, connect.NewRequest(&mgmtv1alpha1.SetPersonalAccountRequest{}))
+	requireNoErrResp(s.T(), resp, err)
+
+	accountId := resp.Msg.GetAccountId()
+	require.NotEmpty(s.T(), accountId)
+}
+
+func (s *IntegrationTestSuite) Test_GetAccounts_NotEmpty() {
+	s.createPersonalAccount(s.unauthUserClient)
+
+	accResp, err := s.unauthUserClient.GetUserAccounts(s.ctx, connect.NewRequest(&mgmtv1alpha1.GetUserAccountsRequest{}))
+	requireNoErrResp(s.T(), accResp, err)
+
+	accounts := accResp.Msg.GetAccounts()
+	require.NotEmpty(s.T(), accounts)
+	require.Len(s.T(), accounts, 1)
+}
+
+func (s *IntegrationTestSuite) Test_IsUserInAccount() {
+	accountId := s.createPersonalAccount(s.unauthUserClient)
+
+	resp, err := s.unauthUserClient.IsUserInAccount(s.ctx, connect.NewRequest(&mgmtv1alpha1.IsUserInAccountRequest{
+		AccountId: accountId,
+	}))
+	requireNoErrResp(s.T(), resp, err)
+	require.True(s.T(), resp.Msg.GetOk())
+
+	resp, err = s.unauthUserClient.IsUserInAccount(s.ctx, connect.NewRequest(&mgmtv1alpha1.IsUserInAccountRequest{
+		AccountId: uuid.NewString(),
+	}))
+	requireNoErrResp(s.T(), resp, err)
+	require.False(s.T(), resp.Msg.GetOk())
+}
+
+func (s *IntegrationTestSuite) Test_CreateTeamAccount_NoAuth() {
+	s.createPersonalAccount(s.unauthUserClient)
+
+	resp, err := s.unauthUserClient.CreateTeamAccount(s.ctx, connect.NewRequest(&mgmtv1alpha1.CreateTeamAccountRequest{Name: "test-name"}))
+	requireErrResp(s.T(), resp, err)
+	requireConnectError(s.T(), err, connect.CodePermissionDenied)
+}
+
+func (s *IntegrationTestSuite) Test_GetTeamAccountMembers_NoAuth_Personal() {
+	accountId := s.createPersonalAccount(s.unauthUserClient)
+
+	resp, err := s.unauthUserClient.GetTeamAccountMembers(s.ctx, connect.NewRequest(&mgmtv1alpha1.GetTeamAccountMembersRequest{AccountId: accountId}))
+	requireErrResp(s.T(), resp, err)
+	requireConnectError(s.T(), err, connect.CodePermissionDenied)
+}
+
+func (s *IntegrationTestSuite) Test_RemoveTeamAccountMember_NoAuth_Personal() {
+	accountId := s.createPersonalAccount(s.unauthUserClient)
+
+	resp, err := s.unauthUserClient.RemoveTeamAccountMember(s.ctx, connect.NewRequest(&mgmtv1alpha1.RemoveTeamAccountMemberRequest{AccountId: accountId, UserId: uuid.NewString()}))
+	requireErrResp(s.T(), resp, err)
+	requireConnectError(s.T(), err, connect.CodePermissionDenied)
+}
+
+func (s *IntegrationTestSuite) Test_InviteUserToTeamAccount_NoAuth_Personal() {
+	accountId := s.createPersonalAccount(s.unauthUserClient)
+
+	resp, err := s.unauthUserClient.InviteUserToTeamAccount(s.ctx, connect.NewRequest(&mgmtv1alpha1.InviteUserToTeamAccountRequest{AccountId: accountId, Email: "test@example.com"}))
+	requireErrResp(s.T(), resp, err)
+	requireConnectError(s.T(), err, connect.CodePermissionDenied)
+}
+
+func (s *IntegrationTestSuite) Test_GetTeamAccountInvites_NoAuth_Personal() {
+	accountId := s.createPersonalAccount(s.unauthUserClient)
+
+	resp, err := s.unauthUserClient.GetTeamAccountInvites(s.ctx, connect.NewRequest(&mgmtv1alpha1.GetTeamAccountInvitesRequest{AccountId: accountId}))
+	requireErrResp(s.T(), resp, err)
+	requireConnectError(s.T(), err, connect.CodePermissionDenied)
+}
+
+func (s *IntegrationTestSuite) Test_RemoveTeamAccountInvite_NoAuth_Personal() {
+	resp, err := s.unauthUserClient.RemoveTeamAccountInvite(s.ctx, connect.NewRequest(&mgmtv1alpha1.RemoveTeamAccountInviteRequest{Id: uuid.NewString()}))
+	requireNoErrResp(s.T(), resp, err)
+}
+
+func (s *IntegrationTestSuite) Test_AcceptTeamAccountInvite_NoAuth_Personal() {
+	resp, err := s.unauthUserClient.AcceptTeamAccountInvite(s.ctx, connect.NewRequest(&mgmtv1alpha1.AcceptTeamAccountInviteRequest{Token: uuid.NewString()}))
+	requireErrResp(s.T(), resp, err)
+	requireConnectError(s.T(), err, connect.CodeUnauthenticated)
+}
+
+func (s *IntegrationTestSuite) Test_GetSystemInformation() {
+	resp, err := s.unauthUserClient.GetSystemInformation(s.ctx, connect.NewRequest(&mgmtv1alpha1.GetSystemInformationRequest{}))
+	requireNoErrResp(s.T(), resp, err)
+}
