--- backend/services/mgmt/v1alpha1/user-account-service/users_integration_auth_test.go
@@ -1,80 +1,0 @@
-package v1alpha1_useraccountservice
-
-import (
-	"connectrpc.com/connect"
-	mgmtv1alpha1 "github.com/nucleuscloud/neosync/backend/gen/go/protos/mgmt/v1alpha1"
-	"github.com/nucleuscloud/neosync/backend/gen/go/protos/mgmt/v1alpha1/mgmtv1alpha1connect"
-	"github.com/nucleuscloud/neosync/backend/internal/authmgmt"
-	"github.com/stretchr/testify/mock"
-	"github.com/stretchr/testify/require"
-)
-
-var (
-	validAuthUser = &authmgmt.User{Name: "foo", Email: "bar", Picture: "baz"}
-)
-
-func (s *IntegrationTestSuite) Test_GetUser_Auth() {
-	client := s.getAuthUserClient("test-user1")
-	userId := s.setUser(client)
-
-	resp, err := client.GetUser(s.ctx, connect.NewRequest(&mgmtv1alpha1.GetUserRequest{}))
-	requireNoErrResp(s.T(), resp, err)
-	require.Equal(s.T(), userId, resp.Msg.GetUserId())
-}
-
-func (s *IntegrationTestSuite) Test_GetUser_Auth_NotFound() {
-	client := s.getAuthUserClient("test-user")
-	resp, err := client.GetUser(s.ctx, connect.NewRequest(&mgmtv1alpha1.GetUserRequest{}))
-	requireErrResp(s.T(), resp, err)
-	requireConnectError(s.T(), err, connect.CodeNotFound)
-}
-
-func (s *IntegrationTestSuite) Test_SetUser_Auth() {
-	client := s.getAuthUserClient("test-user")
-	userId := s.setUser(client)
-	require.NotEmpty(s.T(), userId)
-	require.NotEqual(s.T(), "00000000-0000-0000-0000-000000000000", userId)
-}
-
-func (s *IntegrationTestSuite) Test_CreateTeamAccount_Auth() {
-	client := s.getAuthUserClient("test-user")
-	s.setUser(client)
-
-	resp, err := client.CreateTeamAccount(s.ctx, connect.NewRequest(&mgmtv1alpha1.CreateTeamAccountRequest{Name: "test-name"}))
-	requireNoErrResp(s.T(), resp, err)
-	require.NotEmpty(s.T(), resp.Msg.GetAccountId())
-}
-
-func (s *IntegrationTestSuite) Test_GetTeamAccountMembers_Auth() {
-	client := s.getAuthUserClient("test-user")
-	userId := s.setUser(client)
-	accountId := s.createTeamAccount(client, "test-team")
-
-	s.mockAuthMgmtClient.On("GetUserBySub", mock.Anything, "test-user").
-		Return(validAuthUser, nil)
-
-	resp, err := client.GetTeamAccountMembers(s.ctx, connect.NewRequest(&mgmtv1alpha1.GetTeamAccountMembersRequest{AccountId: accountId}))
-	requireNoErrResp(s.T(), resp, err)
-
-	members := resp.Msg.GetUsers()
-	require.Len(s.T(), members, 1)
-	member := members[0]
-	require.Equal(s.T(), userId, member.GetId())
-	require.Equal(s.T(), validAuthUser.Email, member.GetEmail())
-	require.Equal(s.T(), validAuthUser.Picture, member.GetImage())
-	require.Equal(s.T(), validAuthUser.Name, member.GetName())
-}
-
-func (s *IntegrationTestSuite) setUser(client mgmtv1alpha1connect.UserAccountServiceClient) string {
-	s.T().Helper()
-	resp, err := client.SetUser(s.ctx, connect.NewRequest(&mgmtv1alpha1.SetUserRequest{}))
-	requireNoErrResp(s.T(), resp, err)
-	return resp.Msg.GetUserId()
-}
-
-func (s *IntegrationTestSuite) createTeamAccount(client mgmtv1alpha1connect.UserAccountServiceClient, name string) string {
-	s.T().Helper()
-	resp, err := client.CreateTeamAccount(s.ctx, connect.NewRequest(&mgmtv1alpha1.CreateTeamAccountRequest{Name: name}))
-	requireNoErrResp(s.T(), resp, err)
-	return resp.Msg.AccountId
-}
